backend/requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
gunicorn==23.0.0
PyJWT==2.10.1
bcrypt==4.3.0
requests==2.32.3
pydantic-ai[google]==0.2.14
stripe==12.2.0
pytest==8.3.5
pytest-cov==6.1.1
